refactor: icon types
