+++ projects/ion/src/lib/core/types/icon.ts
@@ -1,0 +1,9 @@
+import { iconsPaths } from '../../icon/svgs/icons';
+
+export type IconType = keyof typeof iconsPaths;
+
+export interface IonIconProps {
+  type: IconType;
+  size?: number;
+  color?: string;
+}
